refactor(config): derive agent defaults from schema
