+++ src/config/types.agent-defaults.ts
@@ -1,11 +1,9 @@
 // Defines agent default configuration types shared by runtime schemas.
-import type { FastMode } from "@openclaw/normalization-core/string-coerce";
-import type { SilentReplyPolicyShape } from "../shared/silent-reply-policy.js";
+import type { z } from "zod";
 import type {
-  AgentModelConfig,
-  AgentToolModelConfig,
   AgentRuntimePolicyConfig,
   AgentSandboxConfig,
+  AgentToolModelConfig,
 } from "./types.agents-shared.js";
 import type {
   BlockStreamingChunkConfig,
@@ -13,6 +11,9 @@ import type {
   HumanDelayConfig,
   TypingMode,
 } from "./types.base.js";
+import type { AgentDefaultsBaseSchema } from "./zod-schema.agent-defaults-base.js";
+
+type SchemaAgentDefaultsConfig = z.input<typeof AgentDefaultsBaseSchema>;
 
 /** Workspace bootstrap-file injection policy for agent system prompts. */
 export type AgentContextInjection = "always" | "continuation-skip" | "never";
@@ -108,7 +109,7 @@ export type AgentContextLimitsConfig = {
   postCompactionMaxChars?: number;
 };
 
-export type AgentDefaultsConfig = {
+export type AgentDefaultsConfig = SchemaAgentDefaultsConfig & {
   /** @deprecated Doctor-only legacy input. */
   imageGenerationModel?: AgentToolModelConfig;
   /** @deprecated Doctor-only legacy input. */
@@ -125,250 +126,37 @@ export type AgentDefaultsConfig = {
   timeFormat?: "auto" | "12" | "24";
   /** @deprecated Doctor-only legacy input. */
   promptOverlays?: { gpt5?: { personality?: "friendly" | "on" | "off" } };
-  /** Global default provider params applied to all models before per-model and per-agent overrides. */
-  params?: Record<string, unknown>;
-  /** Primary model and fallbacks (provider/model). Accepts string or {primary,fallbacks}. */
-  model?: AgentModelConfig;
-  /** Model-selection scope; defaults to the current session. */
-  modelSelectionScope?: ModelSelectionScope;
-  /** Optional lower-cost model for short internal tasks such as generated session titles. */
-  utilityModel?: string;
   /**
    * @deprecated Legacy raw config accepted only by doctor/migration repair.
    * Normal schema parsing rejects this key; use per-model agentRuntime instead.
    */
   agentRuntime?: AgentRuntimePolicyConfig;
-  /** Optional image-capable model and fallbacks (provider/model). Accepts string or {primary,fallbacks}. */
-  imageModel?: AgentToolModelConfig;
-  /** Media-generation model preferences by output modality. */
-  mediaModels?: {
-    image?: AgentToolModelConfig;
-    video?: AgentToolModelConfig;
-    music?: AgentToolModelConfig;
-  };
-  /** Optional voice model and fallbacks (provider/model) for TTS/STT/realtime voice providers. */
-  voiceModel?: AgentToolModelConfig;
-  /** Optional PDF-capable model and fallbacks (provider/model). Accepts string or {primary,fallbacks}. */
-  pdfModel?: AgentToolModelConfig;
-  /** Maximum PDF file size in megabytes (default: 10). */
-  pdfMaxMb?: number;
-  /** Maximum number of PDF pages to process (default: 20). */
-  pdfMaxPages?: number;
-  /** Model catalog with optional aliases (full provider/model keys). */
-  models?: Record<string, AgentModelEntryConfig>;
-  /** Explicit model override policy. Empty or omitted allow permits any model. */
-  modelPolicy?: AgentModelPolicyConfig;
-  /** Agent bootstrap and memory directory; also the working directory when cwd is unset. */
-  workspace?: string;
-  /** Working directory for agent reply runs, separate from workspace bootstrap and memory files. */
-  cwd?: string;
-  /** Optional default allowlist of skills for agents that do not set agents.entries.*.skills. */
-  skills?: string[];
-  /** Silent-reply policy by conversation type. */
-  silentReply?: SilentReplyPolicyShape;
-  /** Optional repository root for system prompt runtime line (overrides auto-detect). */
-  repoRoot?: string;
-  /** Provider-independent prompt overlays applied by model family. */
-  /** Skip bootstrap (BOOTSTRAP.md creation, etc.) for pre-configured deployments. */
-  skipBootstrap?: boolean;
-  /**
-   * List of optional bootstrap filenames to skip writing to the workspace root.
-   * Applies to: SOUL.md, USER.md, IDENTITY.md ("HEARTBEAT.md" is accepted but a no-op).
-   * Required workspace setup such as AGENTS.md still runs.
-   * Example: ["SOUL.md", "USER.md", "IDENTITY.md"]
-   */
-  skipOptionalBootstrapFiles?: OptionalBootstrapFileName[];
-  /**
-   * Controls when workspace bootstrap files (AGENTS.md, SOUL.md, etc.) are
-   * injected into the system prompt:
-   * - always: inject on every turn (default)
-   * - continuation-skip: skip injection on safe continuation turns once the
-   *   transcript already contains a completed assistant turn
-   */
-  contextInjection?: AgentContextInjection;
-  /** Max chars for injected bootstrap files before truncation (default: 20000). */
-  bootstrapMaxChars?: number;
-  /** Max total chars across all injected bootstrap files (default: 150000). */
-  bootstrapTotalMaxChars?: number;
-  /** Experimental agent-default flags. Keep off unless you are intentionally testing a preview surface. */
-  experimental?: {
-    /**
-     * Drop heavyweight non-essential default tools for weaker or smaller local
-     * model backends. Experimental preview only.
-     */
-    localModelLean?: boolean;
-  };
-  /**
-   * Agent-visible bootstrap truncation warning mode:
-   * - off: do not inject warning text
-   * - once: inject once per unique truncation signature
-   * - always: inject on every run with truncation (default)
-   */
-  /**
-   * Optional IANA timezone for model-visible timestamps, prompt context, system events,
-   * and heartbeat active hours. Defaults to the host timezone.
-   */
-  userTimezone?: string;
-  /** Runtime-owned first-turn startup context for bare /new and /reset. */
-  startupContext?: AgentStartupContextConfig;
-  /** Focused context-budget overrides for high-volume injected/read surfaces. */
   contextLimits?: AgentContextLimitsConfig;
-  /** Opt-in: prune old tool results from the LLM context to reduce token usage. */
-  contextPruning?: AgentContextPruningConfig;
-  /** Compaction tuning and pre-compaction memory flush behavior. */
-  compaction?: AgentCompactionConfig;
-  /** Embedded OpenClaw runner hardening and compatibility controls. */
-  embeddedAgent?: {
-    /**
-     * How embedded OpenClaw should trust workspace-local `.openclaw/settings.json`.
-     * - sanitize (default): apply project settings except shellPath/shellCommandPrefix
-     * - ignore: ignore project settings entirely
-     * - trusted: trust project settings as-is
-     */
-    projectSettingsPolicy?: "trusted" | "sanitize" | "ignore";
-    /**
-     * Embedded OpenClaw execution contract:
-     * - default: keep the standard runner behavior
-     * - strict-agentic: enable structured plan tracking and non-visible turn recovery on supported GPT-5 runs
-     */
-    executionContract?: EmbeddedAgentExecutionContract;
-  };
-  /** Default thinking level when no /think directive is present. */
-  thinkingDefault?: AgentThinkingLevel;
-  /** Default fast-mode policy inherited by agent entries that omit it. */
-  fastModeDefault?: FastMode;
-  /** Default verbose level when no /verbose directive is present. */
-  verboseDefault?: "off" | "on" | "full";
-  /**
-   * Detail mode for user-visible tool progress in /verbose and editable progress drafts.
-   * - explain: compact human summary (default)
-   * - raw: include raw command/detail when available
-   */
-  toolProgressDetail?: "explain" | "raw";
-  /** Default reasoning level when no /reasoning directive is present. */
-  reasoningDefault?: "off" | "on" | "stream";
-  /** Default elevated level when no /elevated directive is present. */
-  elevatedDefault?: "off" | "on" | "ask" | "full";
-  /** Default block streaming level when no override is present. */
-  blockStreamingDefault?: "off" | "on";
-  /**
-   * Block streaming boundary:
-   * - "text_end": end of each assistant text content block (before tool calls)
-   * - "message_end": end of the whole assistant message (may include tool blocks)
-   */
-  blockStreamingBreak?: "text_end" | "message_end";
-  /** Soft block chunking for streamed replies (min/max chars, prefer paragraph/newline). */
   blockStreamingChunk?: BlockStreamingChunkConfig;
-  /**
-   * Block reply coalescing (merge streamed chunks before send).
-   * idleMs: wait time before flushing when idle.
-   */
   blockStreamingCoalesce?: BlockStreamingCoalesceConfig;
-  /** Human-like delay between block replies. */
   humanDelay?: HumanDelayConfig;
-  timeoutSeconds?: number;
-  /** Max inbound media size in MB for agent-visible attachments (text note or future image attach). */
-  mediaMaxMb?: number;
-  /**
-   * Max image side length (pixels) when sanitizing base64 image payloads in transcripts/tool results.
-   * Default: 1200.
-   */
-  imageMaxDimensionPx?: number;
-  /**
-   * Image compression/detail preference for image-tool media loading.
-   * Default: auto, which adapts to provider/model limits and image count.
-   */
-  imageQuality?: AgentImageQualityPreference;
-  typingIntervalSeconds?: number;
-  /** Typing indicator start mode (never|instant|thinking|message). */
   typingMode?: TypingMode;
-  /** Periodic background heartbeat runs. */
   heartbeat?: {
-    /** Agent that owns ambient heartbeat runs when no per-agent heartbeat is configured. */
     agentId?: string;
-    /** Heartbeat interval (duration string, default unit: minutes; default: 30m). */
     every?: string;
-    /** Optional active-hours window (local time); heartbeats run only inside this window. */
     activeHours?: {
-      /** Start time (24h, HH:MM). Inclusive. */
       start?: string;
-      /** End time (24h, HH:MM). Exclusive. Use "24:00" for end-of-day. */
       end?: string;
-      /** Timezone for the window ("user", "local", or IANA TZ id). Default: "user". */
       timezone?: string;
     };
-    /** Heartbeat model override (provider/model). */
     model?: string;
-    /** Session key for heartbeat runs ("main" or explicit session key). */
     session?: string;
-    /** Delivery target. Default "owner" uses explicit ownerAllowFrom/allowFrom; "last" may follow groups. */
     target?: string;
-    /** Direct/DM delivery policy. Default: "allow". */
     directPolicy?: "allow" | "block";
-    /** Explicit channel destination; ignored for target "owner" or an unset target. */
     to?: string;
-    /** Optional account id for multi-account channels. */
     accountId?: string;
-    /** Override the heartbeat prompt body. The default treats scratch as monitor prose and directs recurring work to cron jobs. */
     prompt?: string;
-    /** Run timeout in seconds for heartbeat agent turns. Unset uses global timeout or heartbeat cadence capped at 600 seconds. */
     timeoutSeconds?: number;
-    /**
-     * If true, run heartbeat turns with lightweight bootstrap context.
-     * Lightweight mode skips workspace bootstrap files; monitor scratch is
-     * injected by the heartbeat runner either way.
-     */
     lightContext?: boolean;
-    /**
-     * If true, run heartbeat turns in an isolated session with no prior
-     * conversation history. Dramatically reduces per-heartbeat token cost by
-     * avoiding the full session transcript.
-     */
     isolatedSession?: boolean;
   };
-  /** Owner for ambient system-agent/Custodian inference and unscoped operator-read fallbacks. */
-  systemAgent?: {
-    agentId?: string;
-  };
-  /** Upgrade-only owner for the inherited credential store until H2-2 relocates credentials. */
-  authInheritance?: {
-    agentId?: string;
-  };
-  /** Upgrade-only owner for retired main-agent rows and legacy fixed session stores. */
-  sessionStore?: {
-    agentId?: string;
-  };
-  /** Max concurrent agent runs across all conversations. Default: min(16, max(8, available CPU parallelism)). */
-  maxConcurrent?: number;
-  /** Sub-agent defaults (spawned via sessions_spawn). */
-  subagents?: {
-    /** Prompt-only guidance for how strongly the main agent should delegate work. Default: "suggest". */
-    delegationMode?: SubagentDelegationMode;
-    /** Default allowlist of target agent ids for sessions_spawn. Use "*" to allow any configured target. */
-    allowAgents?: string[];
-    /** Max concurrent sub-agent runs (global lane: "subagent"). Default: 8. */
-    maxConcurrent?: number;
-    /** Maximum depth for sessions_spawn chains. Default behavior: 5. */
-    maxSpawnDepth?: number;
-    /** Maximum active children a single requester session may spawn. Default behavior: 5. */
-    maxChildrenPerAgent?: number;
-    /** Auto-archive sub-agent sessions after N minutes (default: 60, set 0 to disable). */
-    archiveAfterMinutes?: number;
-    /** Default model selection for spawned sub-agents (string or {primary,fallbacks}). */
-    model?: AgentModelConfig;
-    /** Default thinking level for spawned sub-agents (e.g. "off", "low", "medium", "high"). */
-    thinking?: string;
-    /** Default run timeout in seconds for spawned sub-agents (0 = no timeout). */
-    runTimeoutSeconds?: number;
-    /** Gateway timeout in ms for sub-agent announce delivery calls (default: 120000). */
-    announceTimeoutMs?: number;
-    /** Require explicit agentId in sessions_spawn (no default same-as-caller). Default: false. */
-    requireAgentId?: boolean;
-  };
-  /** Optional sandbox settings for non-main sessions. */
   sandbox?: AgentSandboxConfig;
 };
-
 export type AgentCompactionMode = "default" | "safeguard";
 export type AgentCompactionPostIndexSyncMode = "off" | "async" | "await";
 export type AgentCompactionIdentifierPolicy = "strict" | "off";
